pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retail-ops-openenv"
version = "0.1.0"
description = "A real-world OpenEnv environment for retail support operations."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "openai>=1.40.0",
  "requests>=2.32.0",
  "httpx>=0.27.0",
  "pytest>=8.3.0",
  "PyYAML>=6.0.2",
  "openenv @ git+https://github.com/meta-pytorch/OpenEnv.git",
]

[tool.setuptools]
packages = ["retail_ops_env", "retail_ops_env.server", "server"]
py-modules = ["client", "models", "tasks", "graders", "validate_submission", "inference"]
